Re #162 kind of writer & initializer for dnd objects + more tests
git-svn-id: https://svn.isis.rl.ac.uk/Horace/branches/OO_IO_operations@1303 1cb955da-4463-431d-88dd-2af2f5550de3

Former-commit-id: c8d545893293de6b00723bc472eae4fc61e777c3

+++ sqw/file_io/@sqw_binfile_common/private/init_sqw_structure_field_by_field_.m
@@ -78,7 +78,7 @@ if obj.num_dim == 'uninitiated' % prototype does not have dimensions in data hea
     obj.num_dim_ = numel(data_header.p_size.field_value);
 end
 if ischar(obj.dnd_dimensions)
-    obj.dnd_dimensions_ = data_header.p_size.field_value;
+    obj.dnd_dimensions_ = double(data_header.p_size.field_value);
 end
 %
 obj.npixels_ = [];
